renamer fil fra index.t
